refactor(geo): merge geo address formatters into country_metadata module
Moves FormatAddress, FormatAddressWithFormat, and AddressComponents from
the standalone geo package into country_metadata, with their tests moved
into country_metadata_test.go. Deletes the geo Go module.

FormatAddressWithFormat now reuses country_metadata's existing data
(GetMetadataInformation().AddressTemplate) instead of geo.go's duplicate
init/cache — so no redundant data-loading code is carried over.

Verified: go build + go vet clean; 457 module tests pass (301 in
country_metadata).
