--- packages/i18nify-go/modules/geo/format_address.go
@@ -1,39 +1,0 @@
-package geo
-
-import (
-	"fmt"
-	"strings"
-)
-
-// FormatAddress formats address components using a caller-supplied template string.
-// Use FormatAddressWithFormat when you have a country code and want the canonical
-// country-specific template from i18nify-data.
-func FormatAddress(template string, components AddressComponents) (string, error) {
-	if strings.TrimSpace(template) == "" {
-		return "", fmt.Errorf("template must be a non-empty string")
-	}
-
-	replacer := strings.NewReplacer(
-		"{name}", components.Name,
-		"{organization}", components.Organization,
-		"{street_address}", components.StreetAddress,
-		"{city}", components.City,
-		"{state}", components.State,
-		"{zip}", components.Zip,
-		"{district}", components.District,
-		"{sorting_code}", components.SortingCode,
-	)
-	substituted := replacer.Replace(template)
-
-	// Split the substituted template into lines and drop any that are blank —
-	// this happens when an optional field (e.g. Organization) was not provided.
-	rawLines := strings.Split(substituted, "\n")
-	formatted := make([]string, 0, len(rawLines))
-	for _, line := range rawLines {
-		if trimmed := strings.TrimSpace(line); trimmed != "" {
-			formatted = append(formatted, trimmed)
-		}
-	}
-
-	return strings.Join(formatted, "\n"), nil
-}
